+++ pyandi/data/dataset.py
@@ -1,6 +1,7 @@
-from os.path import basename, splitext
+from os.path import basename, exists, splitext
 from glob import glob
 import json
+import logging
 import webbrowser
 
 from lxml import etree as ET
@@ -10,6 +11,13 @@ from .activity import ActivitySet
 
 
 class DatasetSet(GenericSet):
+    """Class representing a grouping of ``Dataset`` objects.
+
+    Objects in this grouping can be filtered and iterated over.
+    Queries are only constructed and run when needed, so they
+    can be efficient.
+    """
+
     def __init__(self, data_path, metadata_path, **kwargs):
         super(DatasetSet, self).__init__()
         self._key = 'name'
@@ -41,7 +49,14 @@ class DatasetSet(GenericSet):
 
 
 class Dataset(object):
+    """Class representing an IATI dataset."""
+
     def __init__(self, data_path, metadata_path):
+        """Construct a new Dataset object.
+
+        The file locations of the data and metadata must be specified with
+        the ``data_path`` and ``metadata_path`` arguments.
+        """
         self.data_path = data_path
         self.metadata_path = metadata_path
         self._etree = None
@@ -49,28 +64,43 @@ class Dataset(object):
 
     @property
     def name(self):
+        """Return the name of this dataset, derived from the filename."""
         return splitext(basename(self.data_path))[0]
 
     @property
     def etree(self):
+        """Return the XML of this dataset, as an lxml element tree."""
         if not self._etree:
-            self._etree = ET.parse(self.data_path)
+            try:
+                self._etree = ET.parse(self.data_path)
+            except ET.XMLSyntaxError:
+                logging.warning('Dataset "{}" XML is invalid'.format(
+                    self.name))
+                raise
         return self._etree
 
     @property
     def xml(self):
+        """Return the raw XML of this dataset, as a string."""
         return ET.tostring(self.etree)
 
     def __repr__(self):
         return '<{} ({})>'.format(self.__class__.__name__, self.name)
 
     def show(self):
+        """Open a new browser tab to the iatiregistry.org page
+        for this dataset.
+        """
         url = 'https://iatiregistry.org/dataset/{}'.format(
             self.name)
         webbrowser.open_new_tab(url)
 
     def is_valid(self):
-        # TODO: This currently just checks for valid XML
+        """Check whether the XML in this dataset is valid.
+
+        TODO: This could perform other checks, including validating
+        against a schema.
+        """
         try:
             if self.etree:
                 return True
@@ -80,32 +110,62 @@ class Dataset(object):
 
     @property
     def metadata(self):
-        if not self._metadata:
-            with open(self.metadata_path) as f:
-                self._metadata = json.load(f)
-            self._metadata['extras'] = {x['key']: x['value']
-                                        for x in self.metadata.get('extras')}
+        """Return a dictionary of registry metadata for this dataset."""
+        if self._metadata is None:
+            if exists(self.metadata_path):
+                with open(self.metadata_path) as f:
+                    self._metadata = json.load(f)
+                extras = self.metadata.get('extras')
+                self._metadata['extras'] = {x['key']: x['value']
+                                            for x in extras}
+            else:
+                msg = 'No metadata was found for dataset "{}"'
+                logging.warning(msg.format(self.name))
+                self._metadata = {}
         return self._metadata
 
     @property
     def filetype(self):
-        return self.metadata.get('extras').get('filetype')
+        """Return the filetype according to the metadata
+        (i.e. "activity" or "organisation").
+
+        If it can't be found in the metadata, revert to using
+        the XML root node.
+
+        Return ``None`` if the filetype can't be determined.
+        """
+        try:
+            return self.metadata.get('extras').get('filetype')
+        except AttributeError:
+            pass
+        return {
+            'iati-activities': 'activity',
+            'iati-organisations': 'organisation',
+        }.get(self.root)
 
     @property
     def root(self):
-        roottag = self.etree.getroot().tag
-        if roottag in ['iati-activities', 'iati-organisations']:
-            return roottag
-        return None
+        """Return the name of the XML root node."""
+        return self.etree.getroot().tag
 
     @property
     def version(self):
+        """Return the IATI version according to the XML,
+        or ``None`` if no version specified.
+        """
         try:
             return self.etree.getroot().get('version')
-        except:
+        except ET.XMLSyntaxError:
             pass
+
+        # try:
+        #     return self.metadata.get('extras').get('iati_version')
+        # except AttributeError:
+        #     pass
+
         return None
 
     @property
     def activities(self):
+        """Return an iterator of all activities in this dataset."""
         return ActivitySet([self])