+++ pyandi/utils/abstract.py
@@ -1,8 +1,15 @@
 from itertools import islice
-from .exceptions import OperationError, FilterError, NotFoundError
+from .exceptions import FilterError
 
 
 class GenericSet(object):
+    """Class representing a generic grouping of pyandi objects.
+
+    Objects in this grouping can be filtered and iterated over.
+    Queries are only constructed and run when needed, so they
+    can be efficient.
+    """
+
     def __init__(self):
         self._key = None
         self._filters = []
@@ -10,6 +17,8 @@ class GenericSet(object):
         self._instance_class = None
 
     def where(self, **kwargs):
+        """Return a new set, with the filters provided in ``**kwargs``.
+        """
         for k in kwargs.keys():
             if k.split('__')[0] not in self._filters:
                 raise FilterError('Unknown filter: {}'.format(k))
@@ -17,6 +26,8 @@ class GenericSet(object):
         return self
 
     def filter(self, **kwargs):
+        """Alias of ``where(**kwargs)``.
+        """
         return self.where(**kwargs)
 
     def __getitem__(self, index):
@@ -24,33 +35,56 @@ class GenericSet(object):
             return next(islice(self, index, index + 1))
         except TypeError:
             return list(islice(self, index.start, index.stop, index.step))
+        except StopIteration:
+            pass
+        raise IndexError('index out of range')
 
     def __len__(self):
-        total = 0
-        for x in self:
-            total += 1
-        return total
+        return sum(1 for x in self)
 
     def count(self):
+        """The number of items in this set.
+
+        Equivalent to ``len(self)``.
+        """
         return len(self)
 
     def first(self):
-        for first in self:
-            return first
-        raise NotFoundError
+        """Return the first item in this set.
+
+        Raises an ``IndexError`` if the set contains zero items.
+
+        Equivalent to ``self[0]``.
+        """
+        return self[0]
 
     def all(self):
-        return list(iter(self))
+        """Return a list of all items in this set.
+
+        Equivalent to ``list(self)``.
+        """
+        return list(self)
 
     def get(self, item, default=None):
+        """Return an item from the set, according to the primary key.
+
+        If no matching item is found, ``default`` is returned.
+        """
         if type(item) is self._instance_class:
             item = getattr(item, self._key)
         try:
             return self.find(**{self._key: item})
-        except NotFoundError:
+        except IndexError:
             return default
 
     def find(self, **kwargs):
+        """Return the first matching item from the set, according to the
+        filters provided in ``kwargs``.
+
+        If no matching item is found, an ``IndexError`` is raised.
+
+        ``find(**kwargs)`` is equivalent to ``where(**kwargs).first()``.
+        """
         return self.where(**kwargs).first()
 
 
@@ -76,4 +110,4 @@ class GenericType(object):
                 expr=self.get(),
                 value=value,
             )
-        raise OperationError('Unknown operation: {}'.format(op))
+        raise FilterError('Unknown filter modifier: {}'.format(op))